Renamed old auth button
