+++ packages/commonwealth/client/scripts/views/modals/SessionRevalidationModal/SessionRevalidationModal.tsx
@@ -8,7 +8,7 @@ import WebWalletController from 'controllers/app/web_wallets';
 import useWallets from 'hooks/useWallets';
 import app from 'state';
 import _ from 'underscore';
-import { CWAuthButton } from 'views/components/component_kit/cw_auth_button';
+import { CWAuthButton } from 'views/components/component_kit/CWAuthButtonOld';
 import { CWButton } from 'views/components/component_kit/cw_button';
 import { CWSpinner } from 'views/components/component_kit/cw_spinner';
 import { CWText } from 'views/components/component_kit/cw_text';
